structured and memory bugs fixed

+++ Processor/Processor.c
@@ -26,7 +26,8 @@ Processor *processor_construct()
     Processor *proc = (Processor*)calloc(1, sizeof(Processor));
     proc->stack     = stack_new(START_NUMBER);
     proc->funcs     = stack_new(START_NUMBER);
-    proc->ram       = (double*) calloc(RAM_MEMORY, sizeof(double));
+    //proc->ram       = (double*) calloc(RAM_MEMORY, sizeof(double));
+    proc->ram       = (long long*) calloc(RAM_MEMORY, sizeof(long long));
 
     return proc;
 }
@@ -52,21 +53,33 @@ void start_processing (char *program, size_t length)
 
         char mode = 0;
 
-        double val_earl = 0;
-        double val_last = 0;
+        // double val_earl = 0;
+        // double val_last = 0;
+        long long val_earl = 0;
+        long long val_last = 0;
 
         long long jump  = 0;
 
         program_copy++;
 
+        // printf("value %d\n", (int)val);
+
         switch ((int)val)
         {
-            #include "Commands.h"
+            #include "../Common/Commands.h"
 
             default:
 
                 break;
         }
+
+        // if (val == 1)
+        // {
+        //     printf("length = %u; ", proc->stack->stack->length);
+        //     for (int i = 0; i < proc->stack->stack->length; i++)
+        //         printf("%lld ", proc->stack->stack->buffer[i]);
+        //     printf("\n");
+        // }
     }
 
     processor_destruct(proc);
